Move time-related components to shared module
