+++ client/app/components/shared/time/time_progress.scss
